fix: setup test
